Create NSU_Welfare_CSE499A_Registration_Page
This page is for the users to register into our website. In this page users have to provide their name, valid NSU ID, NSU email, contact number and a password. Then our system will check password validation. Once the form is completed, the information will be saved in our database. The interesting fact is the password which is provided by the user will be totally anonymous. The site managers cannot access it.